+++ app/src/main/java/com/codesoom/assignment/TaskManager.java
@@ -1,0 +1,4 @@
+package com.codesoom.assignment;
+
+public class TaskManager {
+}
