--- app/src/main/java/com/codesoom/assignment/dto/Task.java
@@ -1,23 +1,0 @@
-package com.codesoom.assignment.dto;
-
-public class Task {
-
-    private Long id;
-    private String title;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-}
